+++ transform/src/test/java/com/hellblazer/primeMover/asm/TestAPI.java
@@ -34,9 +34,9 @@ import org.objectweb.asm.util.TraceClassVisitor;
 
 import com.hellblazer.primeMover.TrackingController;
 import com.hellblazer.primeMover.runtime.Framework;
-import com.hellblazer.primeMover.soot.LocalLoader;
 
 import io.github.classgraph.ClassGraph;
+import testClasses.LocalLoader;
 
 /**
  *
