+++ transform/src/test/java/testClasses/ApiUser.java
@@ -1,4 +1,4 @@
-package com.hellblazer.primeMover.soot;
+package testClasses;
 
 public interface ApiUser {
