+++ transform/src/test/java/testClasses/CompositeInterface.java
@@ -1,4 +1,4 @@
-package com.hellblazer.primeMover.soot;
+package testClasses;
 
 public interface CompositeInterface extends Interface1, Interface2 {
