+++ transform/src/test/java/testClasses/EntityThroughSuperclass.java
@@ -17,7 +17,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.hellblazer.primeMover.soot;
+package testClasses;
 
 /**
  *
