+++ transform/src/test/java/testClasses/MockController.java
@@ -1,4 +1,4 @@
-package com.hellblazer.primeMover.soot;
+package testClasses;
 
 import java.util.logging.Logger;
